Share pure CFC label view helpers
