+++ packages/runner/deno.json
@@ -23,6 +23,7 @@
     "./storage/cache": "./src/storage/cache.ts",
     "./storage/cache.deno": "./src/storage/cache.deno.ts",
     "./cfc": "./src/cfc/mod.ts",
+    "./cfc/label-view-core": "./src/cfc/label-view-core.ts",
     "./compilation-cache": "./src/compilation-cache/mod.ts"
   }
 }
