+++ packages/runner/src/cfc/types.ts
@@ -1,6 +1,9 @@
 import type { JSONSchema } from "../builder/types.ts";
 import type { FabricValue, MemorySpace } from "@commonfabric/memory/interface";
 import type { Metadata } from "../storage/interface.ts";
+import type { IFCLabel } from "./label-view-core.ts";
+
+export type { IFCLabel } from "./label-view-core.ts";
 
 export const CFC_STRUCTURAL_PROVENANCE_SETUP_PROJECTION =
   "runtime.setup.result-projection";
@@ -25,11 +28,6 @@ export const isCfcEnforcementMode = (
   CFC_ENFORCEMENT_MODES.includes(input as CfcEnforcementMode);
 
 export const DEFAULT_CFC_ENFORCEMENT_MODE: CfcEnforcementMode = "disabled";
-
-export type IFCLabel = {
-  confidentiality?: unknown[];
-  integrity?: unknown[];
-};
 
 export type CfcSandboxJsonValue =
   | null
